pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repayment"
version = "0.1"
authors = [
    { name = "Eugen Hoppe", email = "rispe_keller0x@icloud.com" }
]
description = "Repayment Plan Generator"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
urls = { "Homepage" = "https://github.com/eugen-hoppe/repayment" }

dependencies = [
    "fpdf2==2.7.9",
    "numpy-financial==1.0.0",
    "python-dateutil==2.9.0.post0"
]

[project.scripts]
stringkeys = "src.repayment:main"

[project.optional-dependencies]

[tool.setuptools]
package-dir = { "" = "src" }
packages = {"find" = { "where" = ["src"], "include" = ["*"] }}
include-package-data = true